requirements.txt
asgiref==3.4.1
boto3==1.21.42
botocore==1.24.42
Django==3.2.5
django-extensions==3.1.3
django-widget-tweaks==1.4.12
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
django-import-export==2.7.1
drf-yasg==1.20.0
pytz==2021.3
sqlparse==0.4.1
mysqlclient
django-crispy-forms==1.12.0
django-bootstrap-breadcrumbs==0.9.2
django-endless-pagination==2.0
pillow
pydantic==1.9.0
phonenumbers==8.12.46
python-environ==0.4.54
elasticsearch==7.8.1
dadata==21.10.1
datetime~=5.1
uwsgi
pymorphy2~=0.9.1
psycopg2
django-jazzmin
django-cors-headers
telebot~=0.0.5

jmespath~=1.0.1
requests~=2.31.0
yookassa~=2.3.6
pandas